# cuda4dnn(DetectionOutput): add fast approximate DetectionOutputOp