--- a/modules/dnn/src/cuda4dnn/csl/tensor.hpp
+++ b/modules/dnn/src/cuda4dnn/csl/tensor.hpp
@@ -780,7 +780,7 @@ namespace cv { namespace dnn { namespace cuda4dnn { namespace csl {
         }
 
         template <class ForwardItr>
-        TensorView(pointer ptr_, ForwardItr start, ForwardItr end) : ptr{ ptr_ } {
+        TensorView(const_pointer ptr_, ForwardItr start, ForwardItr end) : ptr{ ptr_ } {
             CV_Assert(start != end);
             CV_Assert(std::distance(start, end) <= CSL_MAX_TENSOR_RANK);