… modules/dnn/src/cuda/detection_output.cu
… modules/dnn/src/cuda/grid_nms.cu
… modules/dnn/src/cuda4dnn/primitives/detection_output.hpp