Many object detection models can now fully run on the GPU without requiring CPU fallback.

**Approximation in findTopK:**
The confidence values are in the range `[0.0, 1.0]`. The findTopK kernel creates N bins (2048 in this PR). The bins are all stored in the shared memory on the GPU and hence no additional global memory workspace is required. Each block (workgroup) processes one class. The bins are indexed by `(confidence - conf_threshold) *BINS`. The sorting has a precision of 1/N (~`0.0005`). All errors in sorting are off by at most 1/N.

Choosing N is a tradeoff between performance and accuracy. Larger N gives more accurate sorting but is slower.

<cut/>

**Why does ResNet50 Faster RCNN test have low test thresholds?**:
The test fails sometimes. The atomic operations used in approximate count sort are not deterministic. The approximation errors are bounded by `1/N`. Hence, there might be small misordering of boxes.

Faster RCNN's second stage works on the boxes selected by the first DetectionOutput. Even if a slightly different box was selected, there is a chance that the difference is amplified by the second stage network.

The confidence score error is around `0.05` and the IOU is around `0.92`.

**NOTE:** `groupByClasses` default value has been changed to `false` and CUDA backend uses CPU fallback if it is set.

### Pull Request Readiness Checklist

See details at https://github.com/opencv/opencv/wiki/How_to_contribute#making-a-good-pull-request

- [x] I agree to contribute to the project under OpenCV (BSD) License.
- [x] To the best of my knowledge, the proposed patch is not based on a code under GPL or other license that is incompatible with OpenCV
- [x] The PR is proposed to proper branch
- [ ] There is reference to original bug report and related work
- [x] There is accuracy test, performance test and test data in opencv_extra repository, if applicable
      Patch to opencv_extra has the same branch name.
- [ ] The feature is well documented and sample code can be built with the project CMake

```
force_builders=Custom
buildworker:Custom=linux-4
build_image:Custom=ubuntu-cuda:18.04
```