@dkurt 2020-05-15
@YashasSamaga, the thing is that Faster R-CNN has not only final NMS but also intermediate (Proposal layer in Caffe model and intermediate DetectionOutput for TensorFlow). Then the results are used for RoiPooling.