@YashasSamaga 2020-05-15
@dkurt Other FasterRCNN based models and Mask RCNN are all passing. There is exactly one test which is failing. The predictions and references agree on the classes and boxes but the confidence scores are very different.