@YashasSamaga 2020-06-18
The only parameter that is keeping this PR from being ready `groupByClasses`.  This isn't supported yet.

The `groupByClasses` parameter controls how the detection boxes are ordered in the output:
- ordered by scores
- ordered by class and then by score within each class

A lot of models use DetectionOutput but I can't find even one model that actually depends on the ordering of boxes. The final result doesn't change.

InferenceEngine completely ignores `groupByClasses` parameter. The [documentation](https://docs.openvinotoolkit.org/latest/_docs_ops_detection_DetectionOutput_1.html) doesn't say anything about the ordering of boxes.

I can add a check in `supportBackend` to fall back to CPU if `groupByClasses` is set but almost all models have it set. I was thinking of ignoring that parameter just like IE so that models can see performance improvements.