@dkurt 2020-06-18
@YashasSamaga, I also have not seen any model with DetectionOutput layer which groups by classes - all the time it just sorts output by confidence.