@YashasSamaga 2020-06-20
**Device:** RTX  2080 Ti

**Code:** https://gist.github.com/YashasSamaga/48bdb167303e10f4d07b754888ddbdcf

**FP32:**

Model                                | Time (master) | Time (this PR) | Time (this PR + #16900)
------------------------------------ | ------------- | -------------- | -----------------------
Inception v2 Mask RCNN (1024 x 1024) | 50.25ms       | 35.34ms        | 20.39ms 
ResNet50 Faster RCNN (800 x 600)     | 18.54ms       | 9.25ms         | 8.57ms

**FP16:**

Model                                | Time (master) | Time (this PR) | Time (this PR + #16900)
------------------------------------ | ------------- | -------------- | -----------------------
Inception v2 Mask RCNN (1024 x 1024) | 33.10ms       | 17.21ms        | 12.58ms
ResNet50 Faster RCNN (800 x 600)     | 16.91ms       | 7.31ms         | 5.49ms

---

<details>
<summary>Other models and more detailed stats (master)</summary>

```
BVLC AlexNet
[CUDA FP32]
	init >> 32.682ms
	inference >> min = 1.032ms, max = 1.128ms, mean = 1.04105ms, stddev = 0.00941763ms
[CUDA FP16]
	init >> 111.774ms
	inference >> min = 1.455ms, max = 1.654ms, mean = 1.5627ms, stddev = 0.0845136ms


DenseNet 121
[CUDA FP32]
	init >> 55.983ms
	inference >> min = 4.072ms, max = 4.743ms, mean = 4.12198ms, stddev = 0.145124ms
[CUDA FP16]
	init >> 58.287ms
	inference >> min = 4.019ms, max = 5.612ms, mean = 4.04299ms, stddev = 0.158227ms


East Text Detection
[CUDA FP32]
	init >> 123.879ms
	inference >> min = 3.593ms, max = 3.71ms, mean = 3.671ms, stddev = 0.02484ms
[CUDA FP16]
	init >> 126.999ms
	inference >> min = 2.949ms, max = 2.986ms, mean = 2.95566ms, stddev = 0.00447517ms


ENet Cityscapes
[CUDA FP32]
	init >> 25.547ms
	inference >> min = 2.977ms, max = 3.09ms, mean = 2.99362ms, stddev = 0.0132467ms
[CUDA FP16]
	init >> 23.589ms
	inference >> min = 3.964ms, max = 4.246ms, mean = 4.04162ms, stddev = 0.104433ms


FastNeuralStyle Stary Night
[CUDA FP32]
	init >> 20.534ms
	inference >> min = 4.043ms, max = 4.506ms, mean = 4.37642ms, stddev = 0.0681218ms
[CUDA FP16]
	init >> 20.486ms
	inference >> min = 3.16ms, max = 3.252ms, mean = 3.21571ms, stddev = 0.0204613ms


BVLC GoogleNet
[CUDA FP32]
	init >> 16.814ms
	inference >> min = 2.022ms, max = 2.057ms, mean = 2.03271ms, stddev = 0.00609863ms
[CUDA FP16]
	init >> 16.149ms
	inference >> min = 1.954ms, max = 2.144ms, mean = 1.96352ms, stddev = 0.0184839ms


Inception v2 Faster RCNN
[CUDA FP32]
	init >> 100.243ms
	inference >> min = 18.503ms, max = 19.116ms, mean = 18.5449ms, stddev = 0.0597541ms
[CUDA FP16]
	init >> 96.802ms
	inference >> min = 16.818ms, max = 17.963ms, mean = 16.9133ms, stddev = 0.16628ms


Inception v2 Mask RCNN
[CUDA FP32]
	init >> 277.223ms
	inference >> min = 49.912ms, max = 51.025ms, mean = 50.2598ms, stddev = 0.201919ms
[CUDA FP16]
	init >> 256.936ms
	inference >> min = 32.927ms, max = 33.659ms, mean = 33.1076ms, stddev = 0.181882ms


MobileNet SSD
[CUDA FP32]
	init >> 50.061ms
	inference >> min = 36.708ms, max = 38.06ms, mean = 37.0411ms, stddev = 0.254357ms
[CUDA FP16]
	init >> 47.399ms
	inference >> min = 33.824ms, max = 35.947ms, mean = 34.2328ms, stddev = 0.426193ms


MobileNet SSD v1 Coco
[CUDA FP32]
	init >> 19.825ms
	inference >> min = 2.608ms, max = 2.656ms, mean = 2.62962ms, stddev = 0.00897694ms
[CUDA FP16]
	init >> 52.072ms
	inference >> min = 34.638ms, max = 37.079ms, mean = 34.9848ms, stddev = 0.434701ms


MobileNet SSD v2 Coco
[CUDA FP32]
	init >> 44.83ms
	inference >> min = 3.82ms, max = 3.895ms, mean = 3.8377ms, stddev = 0.0100068ms
[CUDA FP16]
	init >> 97.479ms
	inference >> min = 53.487ms, max = 57.548ms, mean = 54.6993ms, stddev = 1.40443ms


OpenCV Face Detector
[CUDA FP32]
	init >> 15.42ms
	inference >> min = 3.942ms, max = 3.987ms, mean = 3.95401ms, stddev = 0.00936686ms
[CUDA FP16]
	init >> 15.385ms
	inference >> min = 4.176ms, max = 4.303ms, mean = 4.19899ms, stddev = 0.0235187ms


OpenFace nn4 small2 v1
[CUDA FP32]
	init >> 20.532ms
	inference >> min = 1.26ms, max = 1.278ms, mean = 1.27012ms, stddev = 0.00274047ms
[CUDA FP16]
	init >> 20.511ms
	inference >> min = 1.273ms, max = 1.294ms, mean = 1.27543ms, stddev = 0.00198341ms


OpenPose pose MPI
[CUDA FP32]
	init >> 61.721ms
	inference >> min = 26.258ms, max = 26.708ms, mean = 26.4663ms, stddev = 0.0563367ms
[CUDA FP16]
	init >> 45.601ms
	inference >> min = 9.45ms, max = 9.626ms, mean = 9.51763ms, stddev = 0.0470375ms


ResNet 50
[CUDA FP32]
	init >> 154.018ms
	inference >> min = 2.541ms, max = 2.585ms, mean = 2.561ms, stddev = 0.00825759ms
[CUDA FP16]
	init >> 153.472ms
	inference >> min = 2.206ms, max = 2.222ms, mean = 2.21263ms, stddev = 0.00239208ms


ResNet50 Faster RCNN
[CUDA FP32]
	init >> 107.402ms
	inference >> min = 23.032ms, max = 24.064ms, mean = 23.5098ms, stddev = 0.26505ms
[CUDA FP16]
	init >> 103.856ms
	inference >> min = 13.121ms, max = 13.511ms, mean = 13.2852ms, stddev = 0.140353ms


ResNet 101
[CUDA FP32]
	init >> 305.266ms
	inference >> min = 4.977ms, max = 5.068ms, mean = 5.02226ms, stddev = 0.0192856ms
[CUDA FP16]
	init >> 296.455ms
	inference >> min = 4.227ms, max = 4.251ms, mean = 4.23403ms, stddev = 0.00516748ms


SqueezeNet v1.1
[CUDA FP32]
	init >> 8.11ms
	inference >> min = 0.449ms, max = 0.461ms, mean = 0.4555ms, stddev = 0.00327321ms
[CUDA FP16]
	init >> 7.457ms
	inference >> min = 0.694ms, max = 0.706ms, mean = 0.69668ms, stddev = 0.00179406ms


Inception v2 Coco
[CUDA FP32]
	init >> 38.786ms
	inference >> min = 5.617ms, max = 5.68ms, mean = 5.63827ms, stddev = 0.0117998ms
[CUDA FP16]
	init >> 38.191ms
	inference >> min = 5.035ms, max = 5.056ms, mean = 5.04418ms, stddev = 0.00617632ms


TensorFlow Inception 5h
[CUDA FP32]
	init >> 20.781ms
	inference >> min = 2.283ms, max = 2.316ms, mean = 2.29716ms, stddev = 0.00560992ms
[CUDA FP16]
	init >> 20.936ms
	inference >> min = 2.229ms, max = 2.291ms, mean = 2.23378ms, stddev = 0.00625305ms


VGG16 SSD
[CUDA FP32]
	init >> 60.73ms
	inference >> min = 17.629ms, max = 17.923ms, mean = 17.7169ms, stddev = 0.0546875ms
[CUDA FP16]
	init >> 67.591ms
	inference >> min = 15.063ms, max = 15.295ms, mean = 15.1127ms, stddev = 0.0396441ms


VGG SSD512
[CUDA FP32]
	init >> 102.367ms
	inference >> min = 26.96ms, max = 27.692ms, mean = 27.2208ms, stddev = 0.161626ms
[CUDA FP16]
	init >> 125.107ms
	inference >> min = 44.824ms, max = 50.284ms, mean = 48.4739ms, stddev = 1.83738ms

YOLO v2
[CUDA FP32]
	init >> 667.816ms
	inference >> min = 8.856ms, max = 10.483ms, mean = 9.03397ms, stddev = 0.346106ms
[CUDA FP16]
	init >> 251.524ms
	inference >> min = 4.619ms, max = 4.699ms, mean = 4.66654ms, stddev = 0.0165728ms


YOLO v3 Tiny
[CUDA FP32]
	init >> 30.95ms
	inference >> min = 1.77ms, max = 1.835ms, mean = 1.77995ms, stddev = 0.0076272ms
[CUDA FP16]
	init >> 31.746ms
	inference >> min = 1.822ms, max = 1.877ms, mean = 1.83245ms, stddev = 0.00640375ms


YOLO v3 Tiny PRN
[CUDA FP32]
	init >> 20.32ms
	inference >> min = 1.575ms, max = 1.623ms, mean = 1.58904ms, stddev = 0.00819964ms
[CUDA FP16]
	init >> 21.103ms
	inference >> min = 1.736ms, max = 1.814ms, mean = 1.75236ms, stddev = 0.012854ms


YOLO v3
[CUDA FP32]
	init >> 338.368ms
	inference >> min = 17.266ms, max = 18.433ms, mean = 17.5562ms, stddev = 0.176949ms
[CUDA FP16]
	init >> 216.782ms
	inference >> min = 7.591ms, max = 7.895ms, mean = 7.83991ms, stddev = 0.0399795ms


YOLO v3 SPP
[CUDA FP32]
	init >> 202.626ms
	inference >> min = 17.598ms, max = 18.951ms, mean = 17.9566ms, stddev = 0.214239ms
[CUDA FP16]
	init >> 195.154ms
	inference >> min = 7.868ms, max = 8.137ms, mean = 8.07337ms, stddev = 0.0334891ms


EfficientNet B0 YOLOv3
[CUDA FP32]
	init >> 38.777ms
	inference >> min = 3.415ms, max = 3.459ms, mean = 3.43736ms, stddev = 0.00799351ms
[CUDA FP16]
	init >> 89.237ms
	inference >> min = 52.545ms, max = 55.501ms, mean = 53.0147ms, stddev = 0.491132ms


YOLO v4
[CUDA FP32]
	init >> 288.501ms
	inference >> min = 18.065ms, max = 19.2ms, mean = 18.3378ms, stddev = 0.157224ms
[CUDA FP16]
	init >> 231.742ms
	inference >> min = 9.745ms, max = 10.032ms, mean = 9.84617ms, stddev = 0.0598817ms
```

</details>

<details>
<summary>Other models and more detailed stats (this PR)</summary>

```
BVLC AlexNet
[CUDA FP32]
	init >> 33.151ms
	inference >> min = 1.03ms, max = 1.053ms, mean = 1.04034ms, stddev = 0.00392148ms
[CUDA FP16]
	init >> 88.726ms
	inference >> min = 1.621ms, max = 1.68ms, mean = 1.63094ms, stddev = 0.00827201ms


DenseNet 121
[CUDA FP32]
	init >> 57.309ms
	inference >> min = 4.076ms, max = 6.856ms, mean = 4.16286ms, stddev = 0.275913ms
[CUDA FP16]
	init >> 59.418ms
	inference >> min = 4.013ms, max = 7.136ms, mean = 4.0519ms, stddev = 0.311209ms


East Text Detection
[CUDA FP32]
	init >> 123.873ms
	inference >> min = 3.594ms, max = 3.705ms, mean = 3.6707ms, stddev = 0.0232947ms
[CUDA FP16]
	init >> 129.16ms
	inference >> min = 2.975ms, max = 2.997ms, mean = 2.98204ms, stddev = 0.00365396ms


ENet Cityscapes
[CUDA FP32]
	init >> 24.896ms
	inference >> min = 2.995ms, max = 3.051ms, mean = 3.00994ms, stddev = 0.00936686ms
[CUDA FP16]
	init >> 22.287ms
	inference >> min = 3.974ms, max = 6.351ms, mean = 4.02333ms, stddev = 0.235467ms


FastNeuralStyle Stary Night
[CUDA FP32]
	init >> 21.507ms
	inference >> min = 4.034ms, max = 4.528ms, mean = 4.37175ms, stddev = 0.0852802ms
[CUDA FP16]
	init >> 21.594ms
	inference >> min = 3.183ms, max = 3.262ms, mean = 3.23938ms, stddev = 0.0172219ms


BVLC GoogleNet
[CUDA FP32]
	init >> 18.884ms
	inference >> min = 2.021ms, max = 2.053ms, mean = 2.03769ms, stddev = 0.00683594ms
[CUDA FP16]
	init >> 16.439ms
	inference >> min = 1.956ms, max = 3.436ms, mean = 1.97854ms, stddev = 0.146519ms


Inception v2 Faster RCNN
[CUDA FP32]
	init >> 92.069ms
	inference >> min = 9.045ms, max = 11.493ms, mean = 9.2525ms, stddev = 0.241762ms
[CUDA FP16]
	init >> 85.876ms
	inference >> min = 7.251ms, max = 7.382ms, mean = 7.31037ms, stddev = 0.0290353ms


Inception v2 Mask RCNN
[CUDA FP32]
	init >> 258.726ms
	inference >> min = 34.943ms, max = 35.649ms, mean = 35.3414ms, stddev = 0.147819ms
[CUDA FP16]
	init >> 258.723ms
	inference >> min = 17.06ms, max = 17.476ms, mean = 17.2136ms, stddev = 0.0843243ms


MobileNet SSD
[CUDA FP32]
	init >> 50.224ms
	inference >> min = 36.383ms, max = 38.706ms, mean = 37.2419ms, stddev = 0.785924ms
[CUDA FP16]
	init >> 50.143ms
	inference >> min = 33.34ms, max = 35.603ms, mean = 33.8624ms, stddev = 0.471735ms


MobileNet SSD v1 Coco
[CUDA FP32]
	init >> 18.354ms
	inference >> min = 1.678ms, max = 1.716ms, mean = 1.69361ms, stddev = 0.0054153ms
[CUDA FP16]
	init >> 51.762ms
	inference >> min = 33.77ms, max = 36.62ms, mean = 34.5305ms, stddev = 0.825172ms


MobileNet SSD v2 Coco
[CUDA FP32]
	init >> 43.079ms
	inference >> min = 2.926ms, max = 2.948ms, mean = 2.93581ms, stddev = 0.00338291ms
[CUDA FP16]
	init >> 96.495ms
	inference >> min = 52.288ms, max = 55.909ms, mean = 53.1842ms, stddev = 1.09509ms


OpenCV Face Detector
[CUDA FP32]
	init >> 12.912ms
	inference >> min = 1.469ms, max = 1.475ms, mean = 1.47183ms, stddev = 0ms
[CUDA FP16]
	init >> 12.792ms
	inference >> min = 1.72ms, max = 1.733ms, mean = 1.72588ms, stddev = 0.00229024ms


OpenFace nn4 small2 v1
[CUDA FP32]
	init >> 19.733ms
	inference >> min = 1.26ms, max = 1.275ms, mean = 1.26828ms, stddev = 0.00343536ms
[CUDA FP16]
	init >> 20.22ms
	inference >> min = 1.275ms, max = 1.38ms, mean = 1.28099ms, stddev = 0.0152583ms


OpenPose pose MPI
[CUDA FP32]
	init >> 61.966ms
	inference >> min = 26.391ms, max = 26.583ms, mean = 26.4763ms, stddev = 0.0358014ms
[CUDA FP16]
	init >> 47.297ms
	inference >> min = 9.414ms, max = 9.607ms, mean = 9.47709ms, stddev = 0.0450491ms


ResNet 50
[CUDA FP32]
	init >> 146.682ms
	inference >> min = 2.543ms, max = 2.58ms, mean = 2.55986ms, stddev = 0.00625305ms
[CUDA FP16]
	init >> 153.241ms
	inference >> min = 2.214ms, max = 2.243ms, mean = 2.22138ms, stddev = 0.00371864ms


ResNet50 Faster RCNN
[CUDA FP32]
	init >> 99.996ms
	inference >> min = 16.987ms, max = 18.217ms, mean = 17.2298ms, stddev = 0.194216ms
[CUDA FP16]
	init >> 95.546ms
	inference >> min = 7.398ms, max = 7.505ms, mean = 7.43676ms, stddev = 0.0247824ms


ResNet 101
[CUDA FP32]
	init >> 307.067ms
	inference >> min = 4.952ms, max = 5.037ms, mean = 4.9966ms, stddev = 0.0205311ms
[CUDA FP16]
	init >> 299.575ms
	inference >> min = 4.231ms, max = 4.262ms, mean = 4.2374ms, stddev = 0.00569429ms


SqueezeNet v1.1
[CUDA FP32]
	init >> 8.101ms
	inference >> min = 0.451ms, max = 0.493ms, mean = 0.45656ms, stddev = 0.00516027ms
[CUDA FP16]
	init >> 7.613ms
	inference >> min = 0.692ms, max = 0.712ms, mean = 0.69529ms, stddev = 0.00221752ms


Inception v2 Coco
[CUDA FP32]
	init >> 37.68ms
	inference >> min = 4.718ms, max = 4.753ms, mean = 4.73546ms, stddev = 0.00662337ms
[CUDA FP16]
	init >> 38.672ms
	inference >> min = 4.151ms, max = 4.18ms, mean = 4.15756ms, stddev = 0.00239208ms


TensorFlow Inception 5h
[CUDA FP32]
	init >> 21.457ms
	inference >> min = 2.265ms, max = 2.295ms, mean = 2.28254ms, stddev = 0.00573601ms
[CUDA FP16]
	init >> 22.073ms
	inference >> min = 2.226ms, max = 2.238ms, mean = 2.23112ms, stddev = 0.00138107ms


VGG16 SSD
[CUDA FP32]
	init >> 54.727ms
	inference >> min = 9.169ms, max = 9.594ms, mean = 9.2823ms, stddev = 0.0818916ms
[CUDA FP16]
	init >> 59.651ms
	inference >> min = 6.048ms, max = 6.199ms, mean = 6.14829ms, stddev = 0.0264935ms


VGG SSD512
[CUDA FP32]
	init >> 88.231ms
	inference >> min = 14.404ms, max = 15.245ms, mean = 14.5701ms, stddev = 0.151288ms
[CUDA FP16]
	init >> 90.921ms
	inference >> min = 11.862ms, max = 12.064ms, mean = 11.931ms, stddev = 0.0447091ms

YOLO v2
[CUDA FP32]
	init >> 223.072ms
	inference >> min = 8.843ms, max = 9.368ms, mean = 8.9732ms, stddev = 0.0951635ms
[CUDA FP16]
	init >> 245.845ms
	inference >> min = 4.612ms, max = 4.709ms, mean = 4.67135ms, stddev = 0.0181651ms


YOLO v3 Tiny
[CUDA FP32]
	init >> 29.589ms
	inference >> min = 1.765ms, max = 1.812ms, mean = 1.77495ms, stddev = 0.00577742ms
[CUDA FP16]
	init >> 30.952ms
	inference >> min = 1.815ms, max = 1.871ms, mean = 1.8241ms, stddev = 0.00785814ms


YOLO v3 Tiny PRN
[CUDA FP32]
	init >> 20.46ms
	inference >> min = 1.577ms, max = 1.609ms, mean = 1.58607ms, stddev = 0.00502716ms
[CUDA FP16]
	init >> 21.545ms
	inference >> min = 1.737ms, max = 1.776ms, mean = 1.74662ms, stddev = 0.00565226ms


YOLO v3
[CUDA FP32]
	init >> 322.953ms
	inference >> min = 17.258ms, max = 18.309ms, mean = 17.5376ms, stddev = 0.183053ms
[CUDA FP16]
	init >> 325.617ms
	inference >> min = 7.748ms, max = 7.858ms, mean = 7.83237ms, stddev = 0.0225245ms


YOLO v3 SPP
[CUDA FP32]
	init >> 316.073ms
	inference >> min = 17.594ms, max = 18.918ms, mean = 17.9141ms, stddev = 0.208756ms
[CUDA FP16]
	init >> 316.447ms
	inference >> min = 7.838ms, max = 8.11ms, mean = 8.0513ms, stddev = 0.03233ms


EfficientNet B0 YOLOv3
[CUDA FP32]
	init >> 39.787ms
	inference >> min = 3.421ms, max = 3.508ms, mean = 3.44294ms, stddev = 0.0117188ms
[CUDA FP16]
	init >> 90.043ms
	inference >> min = 52.441ms, max = 55.622ms, mean = 53.2024ms, stddev = 0.993263ms


YOLO v4
[CUDA FP32]
	init >> 397.104ms
	inference >> min = 18.053ms, max = 19.047ms, mean = 18.3214ms, stddev = 0.145426ms
[CUDA FP16]
	init >> 279.473ms
	inference >> min = 9.733ms, max = 10.078ms, mean = 9.82878ms, stddev = 0.0760466ms
```
</details>


<details>
<summary>Other models and more detailed stats (this PR + #16900 )</summary>

```
BVLC AlexNet
[CUDA FP32]
	init >> 128.848ms
	inference >> min = 1.001ms, max = 1.025ms, mean = 1.00793ms, stddev = 0.00420035ms
[CUDA FP16]
	init >> 113.899ms
	inference >> min = 1.155ms, max = 1.196ms, mean = 1.16575ms, stddev = 0.00618597ms


DenseNet 121
[CUDA FP32]
	init >> 1015.33ms
	inference >> min = 4.374ms, max = 5.278ms, mean = 4.42689ms, stddev = 0.169725ms
[CUDA FP16]
	init >> 239.452ms
	inference >> min = 3.413ms, max = 3.756ms, mean = 3.47972ms, stddev = 0.0852635ms


East Text Detection
[CUDA FP32]
	init >> 2584.08ms
	inference >> min = 3.508ms, max = 3.622ms, mean = 3.57852ms, stddev = 0.0188858ms
[CUDA FP16]
	init >> 301.291ms
	inference >> min = 2.553ms, max = 2.578ms, mean = 2.55934ms, stddev = 0.00420035ms


ENet Cityscapes
[CUDA FP32]
	init >> 436.935ms
	inference >> min = 2.983ms, max = 3.038ms, mean = 3.00095ms, stddev = 0.00910877ms
[CUDA FP16]
	init >> 91.475ms
	inference >> min = 3.247ms, max = 3.456ms, mean = 3.28998ms, stddev = 0.0432894ms


FastNeuralStyle Stary Night
[CUDA FP32]
	init >> 1089.99ms
	inference >> min = 3.985ms, max = 4.117ms, mean = 4.0674ms, stddev = 0.0240797ms
[CUDA FP16]
	init >> 144.082ms
	inference >> min = 2.068ms, max = 4.069ms, mean = 2.15167ms, stddev = 0.193661ms


BVLC GoogleNet
[CUDA FP32]
	init >> 458.161ms
	inference >> min = 1.935ms, max = 1.967ms, mean = 1.94838ms, stddev = 0.00463224ms
[CUDA FP16]
	init >> 121.773ms
	inference >> min = 1.792ms, max = 1.816ms, mean = 1.79729ms, stddev = 0.00390625ms


Inception v2 Faster RCNN
[CUDA FP32]
	init >> 4999.42ms
	inference >> min = 8.328ms, max = 8.711ms, mean = 8.56629ms, stddev = 0.0582018ms
[CUDA FP16]
	init >> 427.71ms
	inference >> min = 5.442ms, max = 5.539ms, mean = 5.49255ms, stddev = 0.019335ms


Inception v2 Mask RCNN
[CUDA FP32]
	init >> 9614.38ms
	inference >> min = 19.989ms, max = 20.755ms, mean = 20.3874ms, stddev = 0.135654ms
[CUDA FP16]
	init >> 959.303ms
	inference >> min = 12.476ms, max = 12.763ms, mean = 12.5844ms, stddev = 0.0532741ms


MobileNet SSD
[CUDA FP32]
	init >> 3370.18ms
	inference >> min = 1.528ms, max = 1.557ms, mean = 1.5448ms, stddev = 0.00609863ms
[CUDA FP16]
	init >> 1585.37ms
	inference >> min = 38.632ms, max = 41.531ms, mean = 39.1035ms, stddev = 0.405197ms


MobileNet SSD v1 Coco
[CUDA FP32]
	init >> 1788.14ms
	inference >> min = 1.65ms, max = 2.151ms, mean = 1.70251ms, stddev = 0.132022ms
[CUDA FP16]
	init >> 834.507ms
	inference >> min = 38.944ms, max = 41.852ms, mean = 39.4362ms, stddev = 0.304988ms


MobileNet SSD v2 Coco
[CUDA FP32]
	init >> 2572.11ms
	inference >> min = 2.876ms, max = 2.912ms, mean = 2.89188ms, stddev = 0.00910877ms
[CUDA FP16]
	init >> 1332.7ms
	inference >> min = 60.562ms, max = 63.856ms, mean = 61.4024ms, stddev = 0.865884ms


OpenCV Face Detector
[CUDA FP32]
	init >> 207.556ms
	inference >> min = 1.444ms, max = 1.469ms, mean = 1.45084ms, stddev = 0.00420035ms
[CUDA FP16]
	init >> 72.149ms
	inference >> min = 1.468ms, max = 1.505ms, mean = 1.47771ms, stddev = 0.00644087ms


OpenFace nn4 small2 v1
[CUDA FP32]
	init >> 241.017ms
	inference >> min = 1.287ms, max = 1.3ms, mean = 1.29362ms, stddev = 0.00300997ms
[CUDA FP16]
	init >> 80.851ms
	inference >> min = 1.153ms, max = 1.213ms, mean = 1.16244ms, stddev = 0.0094366ms


OpenPose pose MPI
[CUDA FP32]
	init >> 5789.7ms
	inference >> min = 21.785ms, max = 22.624ms, mean = 22.2361ms, stddev = 0.177552ms
[CUDA FP16]
	init >> 794.092ms
	inference >> min = 6.729ms, max = 6.858ms, mean = 6.78484ms, stddev = 0.0283035ms


ResNet 50
[CUDA FP32]
	init >> 1160.81ms
	inference >> min = 2.596ms, max = 3.179ms, mean = 2.6704ms, stddev = 0.154479ms
[CUDA FP16]
	init >> 292.971ms
	inference >> min = 2.11ms, max = 2.143ms, mean = 2.11582ms, stddev = 0.00408525ms


ResNet50 Faster RCNN
[CUDA FP32]
	init >> 2499.94ms
	inference >> min = 14.595ms, max = 15.001ms, mean = 14.7541ms, stddev = 0.0854032ms
[CUDA FP16]
	init >> 337.974ms
…